Move methods to reduce lines
