+++ services/app-api/db/options.ts
@@ -2,8 +2,10 @@ import { Choice } from "prompts";
 import {
   getSemiAnnualReportsByState,
   getWorkPlansByState,
+  state,
 } from "../../../tests/seeds/options";
 import { SeedReportShape } from "../../../tests/seeds/types";
+import { generateReportingPeriod } from "./helpers";
 
 export const workPlanChoices = async (): Promise<Choice[]> => {
   try {
@@ -27,4 +29,55 @@ export const semiAnnualReportChoices = async (): Promise<Choice[]> => {
   } catch {
     process.exit();
   }
+};
+
+export const backToMenu = {
+  title: "Back to Menu",
+  value: "back",
+};
+
+export const generateChoices = (
+  type: string,
+  year: number,
+  period: number
+): Choice[] => {
+  const reportingPeriod = generateReportingPeriod(year, period);
+
+  const choices = [
+    {
+      title: `Create base ${type}: ${reportingPeriod}`,
+      value: `create${type}`,
+    },
+    {
+      title: `Create filled ${type}: ${reportingPeriod}`,
+      value: `createFilled${type}`,
+    },
+    {
+      title: `Create submitted ${type}: ${reportingPeriod}`,
+      value: `createSubmitted${type}`,
+    },
+    {
+      title: `Create locked ${type}: ${reportingPeriod}`,
+      value: `createLocked${type}`,
+    },
+    {
+      title: `Create archived ${type}: ${reportingPeriod}`,
+      value: `createArchived${type}`,
+    },
+    { title: `Get ${type} by id`, value: `get${type}ById` },
+    {
+      title: `Get ${type}s by state: ${state}`,
+      value: `get${type}sByState`,
+    },
+    backToMenu,
+  ];
+
+  if (type === "WP") {
+    choices.splice(3, 0, {
+      title: `Create approved ${type}: ${reportingPeriod}`,
+      value: `createApproved${type}`,
+    });
+  }
+
+  return choices;
 };
